refactor: move /tooth/tooth.go to /tooth/toothutils/utils.go
